Move common method to suite file and make tests green
